1/2 half of reading

+++ src/main/java/com/reply/challenges/hashcode2017/codespartans/qualificationround/model/problem/Endpoint.java
@@ -1,7 +1,7 @@
 package com.reply.challenges.hashcode2017.codespartans.qualificationround.model.problem;
 
 import java.util.Arrays;
-import java.util.stream.LongStream;
+import java.util.stream.IntStream;
 
 import lombok.Data;
 
@@ -10,11 +10,11 @@ public class Endpoint {
 
 	private final int id;
 	
-	private final long datacenterLatency;
+	private final int datacenterLatency;
 	
-	private final long[] cacheLatencies;
+	private final int[] cacheLatencies;
 	
-	final LongStream cacheLatencies() {
+	final IntStream cacheLatencies() {
 		return Arrays.stream(this.cacheLatencies);
 	}
 }
