+++ src/main/java/com/reply/challenges/hashcode2017/codespartans/qualificationround/model/problem/Video.java
@@ -7,5 +7,5 @@ public class Video {
 
 	private final int id;
 	
-	private final long size;
+	private final int size;
 }
